Reference

Contents

Reference

Contents

Index

Index

TulipaProfileFitting.find_search_interval

TulipaProfileFitting.find_solution

TulipaProfileFitting.validate_profile

TulipaProfileFitting.find_search_interval — Method

(a, b) = find_search_interval(f)

Returns an interval such that $f(a) × f(b) ≤ 0$. It could be 0 for either endpoint, but it is not positive, ensuring that there is a root in [a, b].

TulipaProfileFitting.find_solution — Method

x = find_solution(P, μ)

Finds a points such that $S(x) = μ$, if possible, where

\[S(x) = \frac{1}{|P|} \sum_{p ∈ P:\ p > 0} p^x.\]

If not possible, return either 0 or 1000, depending on what is most appropriate.

TulipaProfileFitting.validate_profile — Method

validate_profile(P)

Validates if the values of profile P are within 0 and 1. If not, throws an error message and stops the calculation.